[upload] UIString+Extension 생성
